Extract print modal to separate component
